prototype/requirements.txt
Flask
opencv-python
ultralytics
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
google-cloud-firestore
numpy
kivy
torch
torchvision
kaggle
mlflow